fix: satisfy final saju and lovebug gate
